backend/requirements.txt
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
Pygments==2.6.1
python-dotenv==0.14.0
requests==2.24.0
stream-chat==1.3.1
